move _serve to cli/serve
