+++ src/biome/text/cli/serve.py
@@ -1,8 +1,14 @@
+from typing import Any
+from typing import Dict
+
 import click
+import uvicorn
+from allennlp.common.util import sanitize
 from click import Path
+from fastapi import FastAPI
 
 from biome.text import Pipeline
-from biome.text._helpers import _serve
+from biome.text.errors import http_error_handling
 
 
 @click.command()
@@ -34,3 +40,29 @@ def serve(pipeline_path: str, port: int, predictions_dir: str) -> None:
         pipeline.init_prediction_logger(predictions_dir)
 
     return _serve(pipeline, port)
+
+
+def _serve(pipeline: Pipeline, port: int):
+    """Serves an pipeline as rest api"""
+
+    def make_app() -> FastAPI:
+        app = FastAPI()
+
+        @app.post("/predict")
+        async def predict(inputs: Dict[str, Any]):
+            with http_error_handling():
+                return sanitize(pipeline.predict(**inputs))
+
+        @app.get("/_config")
+        async def config():
+            with http_error_handling():
+                return pipeline.config.as_dict()
+
+        @app.get("/_status")
+        async def status():
+            with http_error_handling():
+                return {"ok": True}
+
+        return app
+
+    uvicorn.run(make_app(), host="0.0.0.0", port=port)
